feat: Move fetchRiskToken to components
